chore: rename bases as modules
