+++ src/commands/tree.tsx
@@ -1,5 +1,5 @@
-import { Command } from '@/src/bases/command';
-import { InkCommand } from '@/src/bases/ink-command';
+import { Command } from '@/src/modules/command';
+import { InkCommand } from '@/src/modules/ink-command';
 import { LoadProject } from '@/src/middlewares/load-project';
 import { LoadWorkspace } from '@/src/middlewares/load-workspace';
 import { lazyCurrentWorkspace, type Workspace } from '@/src/project/workspace';
