+++ tests/middlewares/load-project.test.ts
@@ -1,7 +1,7 @@
 import path from 'node:path';
 import yargs from 'yargs';
 
-import { applyMiddlewares } from '@/src/bases/middleware';
+import { applyMiddlewares } from '@/src/modules/middleware';
 import { SpinnerService } from '@/src/commons/spinner.service';
 import { container } from '@/src/inversify.config';
 import { LoadProject } from '@/src/middlewares/load-project';
